+++ lib/capify-cloud/capistrano.rb
@@ -1,44 +1,44 @@
-require File.join(File.dirname(__FILE__), '../capify-ec2')
+require File.join(File.dirname(__FILE__), '../capify-cloud')
 require 'colored'
 
 Capistrano::Configuration.instance(:must_exist).load do  
-  def capify_ec2
-    @capify_ec2 ||= CapifyEc2.new(fetch(:ec2_config, 'config/cloud.yml'))
+  def capify_cloud
+    @capify_cloud ||= CapifyCloud.new(fetch(:cloud_config, 'config/cloud.yml'))
   end
 
-  namespace :ec2 do
+  namespace :cloud do
     
-    desc "Prints out all ec2 instances. index, name, instance_id, size, DNS/IP, region, tags"
+    desc "Prints out all cloud instances. index, name, instance_id, size, DNS/IP, region, tags"
     task :status do
-      capify_ec2.display_instances
+      capify_cloud.display_instances
     end
 
     desc "Deregisters instance from its ELB"
     task :deregister_instance do
       instance_name = variables[:logger].instance_variable_get("@options")[:actions].first
-      capify_ec2.deregister_instance_from_elb(instance_name)
+      capify_cloud.deregister_instance_from_elb(instance_name)
     end
 
     desc "Registers an instance with an ELB."
     task :register_instance do
       instance_name = variables[:logger].instance_variable_get("@options")[:actions].first
       load_balancer_name = variables[:logger].instance_variable_get("@options")[:vars][:loadbalancer]
-      capify_ec2.register_instance_in_elb(instance_name, load_balancer_name)
+      capify_cloud.register_instance_in_elb(instance_name, load_balancer_name)
     end
 
     task :date do
       run "date"
     end
 
-    desc "Prints list of ec2 server names"
+    desc "Prints list of cloud server names"
     task :server_names do
-      puts capify_ec2.server_names.sort
+      puts capify_cloud.server_names.sort
     end
     
     desc "Allows ssh to instance by id. cap ssh <INSTANCE NAME>"
     task :ssh do
       server = variables[:logger].instance_variable_get("@options")[:actions][1]
-      instance = numeric?(server) ? capify_ec2.desired_instances[server.to_i] : capify_ec2.get_instance_by_name(server)
+      instance = numeric?(server) ? capify_cloud.desired_instances[server.to_i] : capify_cloud.get_instance_by_name(server)
       port = ssh_options[:port] || 22 
       command = "ssh -p #{port} #{user}@#{instance.contact_point}"
       puts "Running `#{command}`"
@@ -47,16 +47,16 @@ Capistrano::Configuration.instance(:must_exist).load do
   end
   
   namespace :deploy do
-    before "deploy", "ec2:deregister_instance"
-    after "deploy", "ec2:register_instance"
-    after "deploy:rollback", "ec2:register_instance"
+    before "deploy", "cloud:deregister_instance"
+    after "deploy", "cloud:register_instance"
+    after "deploy:rollback", "cloud:register_instance"
   end
     
-  def ec2_roles(*roles)
+  def cloud_roles(*roles)
     server_name = variables[:logger].instance_variable_get("@options")[:actions].first unless variables[:logger].instance_variable_get("@options")[:actions][1].nil?
     
     if !server_name.nil?
-      named_instance = capify_ec2.get_instance_by_name(server_name)
+      named_instance = capify_cloud.get_instance_by_name(server_name)
   
       task named_instance.name.to_sym do
         remove_default_roles
@@ -66,20 +66,20 @@ Capistrano::Configuration.instance(:must_exist).load do
         end
       end unless named_instance.nil?
     end
-    roles.each {|role| ec2_role(role)}
+    roles.each {|role| cloud_role(role)}
   end
   
-  def ec2_role(role_name_or_hash)
+  def cloud_role(role_name_or_hash)
     role = role_name_or_hash.is_a?(Hash) ? role_name_or_hash : {:name => role_name_or_hash,:options => {}}
     @roles[role[:name]]
     
-    instances = capify_ec2.get_instances_by_role(role[:name])
+    instances = capify_cloud.get_instances_by_role(role[:name])
     if role[:options].delete(:default)
       instances.each do |instance|
         define_role(role, instance)
       end
     end    
-    regions = capify_ec2.determine_regions
+    regions = capify_cloud.determine_regions
     regions.each do |region|
       define_regions(region, role)
     end unless regions.nil?
@@ -92,7 +92,7 @@ Capistrano::Configuration.instance(:must_exist).load do
   def define_regions(region, role)
     instances = []
     @roles.each do |role_name, junk|
-      region_instances = capify_ec2.get_instances_by_region(role_name, region)
+      region_instances = capify_cloud.get_instances_by_region(role_name, region)
       region_instances.each {|instance| instances << instance} unless region_instances.nil?
     end
     task region.to_sym do